# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OmniGest",
        "description": "Documentação da API Omnigest",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://sistema-generico-api.vercel.app/api/api"
        }
    ],
    "paths": {
...
        "/funcionario": {
            "post": {
                "tags": [
                    "Funcionario"
                ],
                "summary": "Cria um novo funcionário",
                "operationId": "59b8de1bfca15211ff442494a65827f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Funcionario"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Funcionário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Funcionario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/funcionario/{id_funcionario}": {
            "get": {
                "tags": [
                    "Funcionario"
                ],
                "summary": "Recupera um funcionário pelo ID",
                "operationId": "8d9229308b5fd388b94c2437bfb9c64b",
                "parameters": [
                    {
                        "name": "id_funcionario",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Funcionário recuperado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Funcionario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Funcionario"
                ],
                "summary": "Atualiza um funcionário",
                "operationId": "c2d582e287ec7742d656b46cd9ccd912",
                "parameters": [
                    {
                        "name": "id_funcionario",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Funcionario"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Funcionário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Funcionario"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação ou funcionário não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Funcionario"
                ],
                "summary": "Deleta um funcionário",
                "operationId": "15de65220d4234937303e3be39bcf196",
                "parameters": [
                    {
                        "name": "id_funcionario",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Funcionário deletado com sucesso"
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Funcionario": {
                "properties": {
                    "id_funcionario_tfu": {
                        "type": "integer"
                    },
                    "id_empresa_tfu": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "id_funcionario_cargo_tfu": {
                        "description": "Chave estrangeira referenciando o cargo do funcionário",
                        "type": "integer"
                    },
                    "desc_funcionario_tfu": {
                        "type": "string"
                    },
                    "telefone_funcionario_tfu": {
                        "type": "string"
                    },
                    "documento_funcionario_tfu": {
                        "type": "string"
                    },
                    "endereco_funcionario_tfu": {
                        "type": "string"
                    },
                    "is_ativo_tfu": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id_centro_custo_tfu": {
                        "description": "Chave estrangeira referenciando o centro de custo",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}